PRINT_ELEMENTS_LIMIT is now private
