+++ src/arrayformat.rs
@@ -15,7 +15,7 @@ use super::{
 };
 use crate::dimension::IntoDimension;
 
-pub const PRINT_ELEMENTS_LIMIT: Ix = 3;
+const PRINT_ELEMENTS_LIMIT: Ix = 3;
 
 fn format_array_v2<A, S, D, F>(view: &ArrayBase<S, D>,
                                f: &mut fmt::Formatter,
@@ -289,5 +289,132 @@ impl<'a, A: fmt::Binary, S, D: Dimension> fmt::Binary for ArrayBase<S, D>
 {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         format_array_v2(self, f, <_>::fmt, PRINT_ELEMENTS_LIMIT)
+    }
+}
+
+#[cfg(test)]
+mod formatting_with_omit {
+    use crate::prelude::*;
+    use super::*;
+
+    fn print_output_diff(expected: &str, actual: &str) {
+        println!("Expected output:\n{}\nActual output:\n{}", expected, actual);
+    }
+
+    #[test]
+    fn dim_1() {
+        let overflow: usize = 5;
+        let a = Array1::from_elem((PRINT_ELEMENTS_LIMIT * 2 + overflow, ), 1);
+        let mut expected_output = String::from("[");
+        a.iter()
+            .take(PRINT_ELEMENTS_LIMIT)
+            .for_each(|elem| { expected_output.push_str(format!("{}, ", elem).as_str()) });
+        expected_output.push_str("...");
+        a.iter()
+            .skip(PRINT_ELEMENTS_LIMIT + overflow)
+            .for_each(|elem| { expected_output.push_str(format!(", {}", elem).as_str()) });
+        expected_output.push(']');
+        let actual_output = format!("{}", a);
+
+        print_output_diff(&expected_output, &actual_output);
+        assert_eq!(actual_output, expected_output);
+    }
+
+    #[test]
+    fn dim_2_last_axis_overflow() {
+        let overflow: usize = 3;
+        let a = Array2::from_elem((PRINT_ELEMENTS_LIMIT, PRINT_ELEMENTS_LIMIT * 2 + overflow), 1);
+        let mut expected_output = String::from("[");
+
+        for i in 0..PRINT_ELEMENTS_LIMIT {
+            expected_output.push_str(format!("[{}", a[(i, 0)]).as_str());
+            for j in 1..PRINT_ELEMENTS_LIMIT {
+                expected_output.push_str(format!(", {}", a[(i, j)]).as_str());
+            }
+            expected_output.push_str(", ...");
+            for j in PRINT_ELEMENTS_LIMIT + overflow..PRINT_ELEMENTS_LIMIT * 2 + overflow {
+                expected_output.push_str(format!(", {}", a[(i, j)]).as_str());
+            }
+            expected_output.push_str(if i < PRINT_ELEMENTS_LIMIT - 1 { "],\n " } else { "]" });
+        }
+        expected_output.push(']');
+        let actual_output = format!("{}", a);
+
+        print_output_diff(&expected_output, &actual_output);
+        assert_eq!(actual_output, expected_output);
+    }
+
+    #[test]
+    fn dim_2_non_last_axis_overflow() {
+        let overflow: usize = 5;
+        let a = Array2::from_elem((PRINT_ELEMENTS_LIMIT * 2 + overflow, PRINT_ELEMENTS_LIMIT), 1);
+        let mut expected_output = String::from("[");
+
+        for i in 0..PRINT_ELEMENTS_LIMIT {
+            expected_output.push_str(format!("[{}", a[(i, 0)]).as_str());
+            for j in 1..PRINT_ELEMENTS_LIMIT {
+                expected_output.push_str(format!(", {}", a[(i, j)]).as_str());
+            }
+            expected_output.push_str("],\n ");
+        }
+        expected_output.push_str("...,\n ");
+        for i in PRINT_ELEMENTS_LIMIT + overflow..PRINT_ELEMENTS_LIMIT * 2 + overflow {
+            expected_output.push_str(format!("[{}", a[(i, 0)]).as_str());
+            for j in 1..PRINT_ELEMENTS_LIMIT {
+                expected_output.push_str(format!(", {}", a[(i, j)]).as_str());
+            }
+            expected_output.push_str(if i == PRINT_ELEMENTS_LIMIT * 2 + overflow - 1 {
+                "]"
+            } else {
+                "],\n "
+            });
+        }
+        expected_output.push(']');
+        let actual_output = format!("{}", a);
+
+        print_output_diff(&expected_output, &actual_output);
+        assert_eq!(actual_output, expected_output);
+    }
+
+    #[test]
+    fn dim_2_multi_directional_overflow() {
+        let overflow: usize = 5;
+        let a = Array2::from_elem(
+            (PRINT_ELEMENTS_LIMIT * 2 + overflow, PRINT_ELEMENTS_LIMIT * 2 + overflow), 1
+        );
+        let mut expected_output = String::from("[");
+
+        for i in 0..PRINT_ELEMENTS_LIMIT {
+            expected_output.push_str(format!("[{}", a[(i, 0)]).as_str());
+            for j in 1..PRINT_ELEMENTS_LIMIT {
+                expected_output.push_str(format!(", {}", a[(i, j)]).as_str());
+            }
+            expected_output.push_str(", ...");
+            for j in PRINT_ELEMENTS_LIMIT + overflow..PRINT_ELEMENTS_LIMIT * 2 + overflow {
+                expected_output.push_str(format!(", {}", a[(i, j)]).as_str());
+            }
+            expected_output.push_str("],\n ");
+        }
+        expected_output.push_str("...,\n ");
+        for i in PRINT_ELEMENTS_LIMIT + overflow..PRINT_ELEMENTS_LIMIT * 2 + overflow {
+            expected_output.push_str(format!("[{}", a[(i, 0)]).as_str());
+            for j in 1..PRINT_ELEMENTS_LIMIT {
+                expected_output.push_str(format!(", {}", a[(i, j)]).as_str());
+            }
+            expected_output.push_str(", ...");
+            for j in PRINT_ELEMENTS_LIMIT + overflow..PRINT_ELEMENTS_LIMIT * 2 + overflow {
+                expected_output.push_str(format!(", {}", a[(i, j)]).as_str());
+            }
+            expected_output.push_str(if i == PRINT_ELEMENTS_LIMIT * 2 + overflow - 1 {
+                "]"
+            } else {
+                "],\n "
+            });
+        }
+        expected_output.push(']');
+        let actual_output = format!("{}", a);
+
+        print_output_diff(&expected_output, &actual_output);
+        assert_eq!(actual_output, expected_output);
     }
 }
